+++ NoneBot/src/plugins/siyuan/config.py
@@ -25,7 +25,8 @@ class SiyuanConfig(BaseModel):
     siyuan_pgp_comment: str = "SiYuan Bot"  # PGP 密钥用户备注
     siyuan_pgp_email: str = ""  # PGP 密钥用户邮箱
 
-    siyuan_assets_upload_url: str = "https://ld246.com/api/assets/upload"  # 思源云上传地址
+    siyuan_assets_add_url: str = "https://ld246.com/apis/siyuan/inbox/addCloudShorthand"  # 云收集箱-内容添加地址
+    siyuan_assets_upload_url: str = "https://ld246.com/apis/siyuan/upload"  # 云收集箱-资源文件上传地址
 
     # 业务类型
     siyuan_assets_upload_biz_type_key: str = "Biz-Type"
@@ -37,6 +38,6 @@ class SiyuanConfig(BaseModel):
 
     # 用户代理
     siyuan_assets_upload_user_agent_key: str = "User-Agent"
-    siyuan_assets_upload_user_agent_value: str = "SiYuan/0.0.0 SiYuan-QQ-bot/0.0.0"
+    siyuan_assets_upload_user_agent_value: str = "SiYuan/0.0.0"
 
     siyuan_data_file_name: str = "data.json"  # 数据文件名