+++ NoneBot/src/plugins/siyuan/plugins/account/user.py
@@ -13,6 +13,8 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from email import message
+from functools import partial
 from urllib.parse import (
     ParseResult,
     urlparse,
@@ -26,6 +28,7 @@ import nonebot.adapters.qq.models as models
 
 from ... import data
 from ...data import InboxMode
+from ...reply import reply
 
 current_user = on_command(
     cmd="user",
@@ -69,6 +72,12 @@ async def _(
     bot: ob.Bot | qq.Bot,
     event: ob.MessageEvent | qq.MessageEvent,
 ):
+    reply_ = partial(
+        reply,
+        bot=bot,
+        event=event,
+        matcher=current_user,
+    )
     user_id = event.get_user_id()
     account = data.getAccount(user_id)
     match account.inbox.mode:
@@ -84,42 +93,38 @@ async def _(
         if len(account.service.baseURI) == 0
         else desensitizeURI(urlparse(account.service.baseURI))
     )
-    reply = "\n".join(
-        [
-            f"- 当前用户 (id): {account.id}",
-            f"- 收集箱 (inbox)",
-            f"  - 是否已启用 (enable): {account.inbox.enable}",
-            f"  - 默认模式 (mode): {mode}",
-            f"- 云服务 (cloud):",
-            f"  - 链滴 API 令牌 (token): {desensitizeString(account.cloud.token)}",
-            f"- 内核服务 (service):",
-            f"  - 内核服务地址 (baseURI): {baseURI}",
-            f"  - 内核服务令牌 (token): {desensitizeString(account.service.token)}",
-            f"  - 资源文件目录 (assets): {desensitizeString(account.service.assets)}",
-            f"  - 笔记本 ID (notebook): {desensitizeString(account.service.notebook)}",
-        ]
-    )
-    match bot:
-        case ob.Bot():
-            await current_user.finish(ob.MessageSegment.text(reply))
-        case qq.Bot():
-            # Markdown 消息模板需要申请
-            # await current_user.finish(qq.MessageSegment.markdown(reply))
-            match event:
-                case qq.QQMessageEvent():  # 群聊/单聊
-                    await current_user.finish(ob.MessageSegment.text(reply))
-                case qq.GuildMessageEvent():  # 频道/私信
-                    await current_user.finish(
-                        qq.MessageSegment.embed(
-                            qq.message.MessageEmbed(
-                                title=event.author.username,
-                                prompt=f"用户信息 [{event.author.username}]",
-                                description=f"用户 [@{event.author.username}] 绑定的信息",
-                                # thumbnail={
-                                #     "url": event.author.avatar,
-                                # },
-                                ## fields 中的 name 不能为空字符串, 否则消息卡片会显示为空白
-                                fields=[models.MessageEmbedField(name=name) for name in reply.split("\n")],
-                            )
-                        )
+    lines = [
+        f"- 当前用户 (id): {account.id}",
+        f"- 收集箱 (inbox)",
+        f"  - 是否已启用 (enable): {account.inbox.enable}",
+        f"  - 默认模式 (mode): {mode}",
+        f"- 云服务 (cloud):",
+        f"  - 链滴 API 令牌 (token): {desensitizeString(account.cloud.token)}",
+        f"- 内核服务 (service):",
+        f"  - 内核服务地址 (baseURI): {baseURI}",
+        f"  - 内核服务令牌 (token): {desensitizeString(account.service.token)}",
+        f"  - 资源文件目录 (assets): {account.service.assets}",
+        f"  - 笔记本 ID (notebook): {desensitizeString(account.service.notebook)}",
+    ]
+
+    match event:
+        case qq.QQMessageEvent():  # 群聊/单聊
+            await reply_(qq.MessageSegment.text("\n".join(lines)))
+        case qq.GuildMessageEvent():  # 频道/私信
+            ## Markdown 消息模板需要申请
+            # await reply_(message=qq.MessageSegment.markdown(message))
+            await reply_(
+                message=qq.MessageSegment.embed(
+                    qq.message.MessageEmbed(
+                        title=event.author.username,
+                        prompt=f"用户信息 [{event.author.username}]",
+                        description=f"用户 [@{event.author.username}] 绑定的信息",
+                        # thumbnail={
+                        #     "url": event.author.avatar,
+                        # },
+                        ## fields 中的 name 不能为空字符串, 否则消息卡片会显示为空白
+                        fields=[models.MessageEmbedField(name=name) for name in lines],
                     )
+                ),
+                reference=False,
+            )