+++ NoneBot/src/plugins/siyuan/plugins/inbox/__init__.py
@@ -13,17 +13,21 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from functools import partial
 from nonebot import on_message
 from nonebot.plugin import PluginMetadata
 import nonebot.adapters.onebot.v11 as ob
 import nonebot.adapters.qq as qq
 
 from ... import data
+from ...client import Client
 from ...data import InboxMode
+from ...reply import reply
 from . import (
     middleware,
     settings,
 )
+from .transfer import Transfer
 
 __plugin_meta__ = PluginMetadata(
     name="inbox",
@@ -42,16 +46,45 @@ async def _(
     bot: ob.Bot | qq.Bot,
     event: ob.MessageEvent | qq.MessageEvent,
 ):
+    reply_ = partial(
+        reply,
+        bot=bot,
+        event=event,
+        matcher=inbox_default,
+    )
+
     # 判断当前默认收集箱方案
     user_id = event.get_user_id()
     account = data.getAccount(user_id)
     if account.inbox.enable:
-        match account.inbox.mode:
-            case InboxMode.none:
-                pass
-            case InboxMode.cloud:
-                pass
-            case InboxMode.service:
-                pass
-    print(event.json())
-    # TODO: 解析消息
+        client = Client.new(account)
+        transfer = Transfer(client)
+        content: str
+
+        # 解析消息
+        try:
+            content = await transfer.msg2md(
+                mode=account.inbox.mode,
+                message=event.get_message(),
+                event=event,
+            )
+        except Exception as e:
+            await reply_("解析消息异常")
+            # await reply_(f"解析消息异常：\n{e}")
+
+        # 上传收集箱内容
+        inbox_mode: str
+        try:
+            match account.inbox.mode:
+                case InboxMode.none:
+                    inbox_mode = "未设置默认收集箱"
+                case InboxMode.cloud:
+                    await client.addCloudShorthand(content=content)
+                    inbox_mode = "云收集箱"
+                case InboxMode.service:
+                    inbox_mode = "思源收集箱"
+        except Exception as e:
+            await reply_(f"上传收集箱异常：\n{e}")
+        await reply_(f"已加入收集箱: {inbox_mode}")
+    else:
+        await reply_("收集箱未启用")